Huge color uniquness audit.
To do this I added logging to block.cc in each of the 3 places it
sets a color. The logging output the hex code for the color and
which of the three places it was (block, variant or auto) as well
as the name of the thing being set. This log output was then sifted
to get the unique list of colors being set, each color was then
searched in the list of all the colors set to see how many times it
was used. Anything used more than once was adjusted to no longer
have conflicts.