DimensionData_LevelDB chunk bounds initialization.
We don't need the limits passed in to DimensionData, nor do we need to
initialize the bounds to 0, 0.
By setting the max to INT32_MIN and the min to INT32_MAX, they'll get
updated to the location of the first chunk passed in.

This simplifies the (unmerged) change by wgebczyk, and also removes the
issue where the map will get extended to include 0, 0 even when it isn't
present in the current world.