Update the extent when changing layers.
This was potentially an issue before this change, but was unlikely to
come up in practice. When using tiles, and we switch dimensions, the
full extent of the layer wasn't being updated. Usually wasn't an issue
as the overworld map is usually the largest and that's loaded first.
It's a bigger issue with the limits, because if you limit the overworld
only, the other dimensions (while generated) can't be seen through
OpenLayers.