adapt block and item to the "new" non-id based world.
We still use ID extensively internally to cross reference things
so we need to have an ID for all the things, but Mojang have stopped
issuing IDs to new blocks and items, they only get unames now. So we
assign an internal auto-incrementing id to all the blocks and items
if there isn't a known legacy one. This id should only be used inside
the mapping software, not exported outside as it is volatile.

The storage of these ids now uses a variable sized container (std::vector)
to map the ID to the block/item object, as well as an std::unordered_map to map
the various unames to the object. (note that a single block should only appear
in the allThingsById map once, whereas it can well occur in the unameThingMap
more than once.)