src/world: fix buffer overflow parsing VILLAGE tags
Evidently the VILLAGE tag has changed in later versions of Bedrock to
include more information before the unique VID. Make the code more
secure by searching backwards for the "_" delimiters and only copying
the villiger ID if it contains the key. Avoid the additional copy of
rectype by checking for INFO directly from the found delimiter.

By correctly parsing these new villager tags we also need to extend
the villager parsing code to check for errors and try the two forms of
the key I've seen (there may be more). It is a fairly clunkly solution
but should prevent future breakage.